fix: update organize folder componentes modal
