--- src/components/Modal/Modal.types.ts
@@ -1,8 +1,0 @@
-export type ModalProps = {
-  openModalComponent?: React.ReactElement;
-  type: React.ReactNode;
-  children: React.ReactNode;
-  closeModalComponent?: React.ReactElement;
-  open?: boolean;
-  onOpenChange?: (open: boolean) => void;
-};
