--- src/components/Modal/ModalRoot/index.tsx
@@ -1,7 +1,0 @@
-import ModalContent from "./ModalContent";
-import ModalRoot from "./ModalRoot";
-
-export const Modal = {
-  Root: ModalRoot,
-  Content: ModalContent,
-};
