--- src/components/Modal/index.tsx
@@ -1,2 +1,0 @@
-export { default } from "./Modal";
-export type { ModalProps } from "./Modal.types";
